feat: move TestxServer in a seprate file
